chore: resolve merge conflicts and sync dogfood baseline (v0.12.13)
Merged latest main branch changes (v0.12.13) and updated
the dogfood baseline to resolve lint violations in the new documents.
