+++ .github/workflows/release-on-merge.yml
@@ -1,0 +1,61 @@
+name: Release on Merge
+
+# Fires when a release/vX.Y.Z PR is merged to main.
+# Creates a GPG-signed annotated tag; the existing release.yml picks it up via push: tags: v*.
+#
+# Required secrets:
+#   RELEASE_GPG_KEY        — ASCII-armored GPG private key (export: gpg --armor --export-secret-keys KEY_ID)
+#   RELEASE_GPG_PASSPHRASE — passphrase for the key (empty string if none)
+
+on:
+  pull_request:
+    types: [closed]
+    branches: [main, master]
+
+concurrency:
+  group: release-on-merge-${{ github.sha }}
+  cancel-in-progress: false
+
+permissions:
+  contents: write
+
+jobs:
+  auto-tag:
+    name: Create release tag
+    if: >
+      github.event.pull_request.merged == true &&
+      startsWith(github.event.pull_request.head.ref, 'release/v')
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout main
+        uses: actions/checkout@v5
+        with:
+          ref: main
+          fetch-depth: 0
+
+      - name: Extract version from Cargo.toml
+        id: ver
+        run: |
+          bare=$(grep '^version = ' Cargo.toml | head -1 | sed 's/.*"\(.*\)".*/\1/')
+          echo "version=v${bare}" >> "$GITHUB_OUTPUT"
+          echo "Resolved release version: v${bare}"
+
+      - name: Assert tag is safe to create
+        run: scripts/release/assert-tag-safe.sh "${{ steps.ver.outputs.version }}"
+
+      - name: Import GPG signing key
+        uses: crazy-max/ghaction-import-gpg@v6
+        with:
+          gpg_private_key: ${{ secrets.RELEASE_GPG_KEY }}
+          passphrase: ${{ secrets.RELEASE_GPG_PASSPHRASE }}
+          git_user_signingkey: true
+          git_tag_gpgsign: true
+          git_config_global: true
+
+      - name: Create signed annotated tag
+        run: |
+          git tag -s "${{ steps.ver.outputs.version }}" \
+            -m "Release ${{ steps.ver.outputs.version }}"
+          git push origin "${{ steps.ver.outputs.version }}"
+          echo "Tag ${{ steps.ver.outputs.version }} pushed. release.yml will pick it up automatically."
