+++ .github/workflows/release.yml
@@ -136,7 +136,9 @@ jobs:
           fi
 
       - name: Publish crate
-        if: github.event_name == 'workflow_dispatch' && inputs.publish_crate == true
+        if: >
+          (github.event_name == 'push' && startsWith(github.ref, 'refs/tags/')) ||
+          (github.event_name == 'workflow_dispatch' && inputs.publish_crate == true)
         env:
           CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
         run: |
