+++ openspec/changes/archive/2026-04-27-v0-12-12-quality-and-performance-hardening/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-26
