--- openspec/changes/v0-12-12-quality-and-performance-hardening/public-confidence-score.json
@@ -1,44 +1,0 @@
-{
-  "version": "v0.12.12",
-  "status": "pending",
-  "score": {
-    "total": 0,
-    "threshold": 90,
-    "categories": [
-      {
-        "name": "External corpus confidence",
-        "score": 0,
-        "max": 30,
-        "evidence": []
-      },
-      {
-        "name": "Precision regression",
-        "score": 0,
-        "max": 25,
-        "evidence": []
-      },
-      {
-        "name": "Command convergence",
-        "score": 0,
-        "max": 20,
-        "evidence": []
-      },
-      {
-        "name": "Performance stability",
-        "score": 0,
-        "max": 15,
-        "evidence": []
-      },
-      {
-        "name": "Release reproducibility",
-        "score": 0,
-        "max": 10,
-        "evidence": []
-      }
-    ]
-  },
-  "technical_hard_blockers": [],
-  "known_limitations": [],
-  "non_blocking_follow_ups": [],
-  "verification": {}
-}
