+++ openspec/changes/v0-16-0-config-schema-and-editor-integration/.openspec.yaml
