+++ scripts/ci/coverage-baseline.txt
@@ -1,1 +1,1 @@
-623
+880
