+++ src/cli/mod.rs
@@ -1,0 +1,7 @@
+pub mod args;
+pub(crate) mod input;
+pub(crate) mod reporter;
+pub(crate) mod workflow;
+
+pub use args::{parse_args, Cli, Command, ConfigCommand, OutputFormat};
+pub use workflow::{run, run_from_env};
