+++ src/rules/markdown/rules/md046.rs
@@ -46,7 +46,25 @@ impl MarkdownRule for CodeBlockStyleRule {
 }
 
 fn is_indented_code_line(ctx: &DocumentContext<'_>, line_index: usize, line: &str) -> bool {
-    !ctx.is_code_line(line_index) && line.starts_with("    ") && !line.trim().is_empty()
+    if ctx.is_code_line(line_index) || !line.starts_with("    ") || line.trim().is_empty() {
+        return false;
+    }
+    !is_list_marker_line(&line[4..])
+}
+
+fn is_list_marker_line(s: &str) -> bool {
+    let s = s.trim_start();
+    if s.starts_with("- ")
+        || s.starts_with("* ")
+        || s.starts_with("+ ")
+        || s == "-"
+        || s == "*"
+        || s == "+"
+    {
+        return true;
+    }
+    let num_end = s.find(|c: char| !c.is_ascii_digit()).unwrap_or(0);
+    num_end > 0 && (s[num_end..].starts_with(". ") || s[num_end..].starts_with(") "))
 }
 
 #[cfg(test)]
@@ -56,6 +74,15 @@ mod tests {
     #[test]
     fn ignores_indented_lines_inside_fenced_diagrams() {
         let content = "# Title\n\n```mermaid\ngraph TD\n    A --> B\n```\n";
+        let results = lint(content, &LintOptions::default()).expect("lint runs");
+
+        assert!(results.iter().all(|result| result.rule_id != "MD046"));
+    }
+
+    #[test]
+    fn ignores_indented_list_items_as_code_block() {
+        // 4-space indented list items must not be treated as indented code blocks
+        let content = "```rust\nlet x = 1;\n```\n\n- item\n    - nested\n    - nested2\n";
         let results = lint(content, &LintOptions::default()).expect("lint runs");
 
         assert!(results.iter().all(|result| result.rule_id != "MD046"));
